add notifications_settings #33
