+++ app/build.gradle
@@ -10,6 +10,7 @@ android {
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
         release {
@@ -26,6 +27,7 @@ android {
         abortOnError false
         disable 'InvalidPackage'
     }
+
 
 }
 
@@ -54,5 +56,4 @@ dependencies {
     testCompile 'junit:junit:4.12'
 //firebase job dispatcher
     compile 'com.firebase:firebase-jobdispatcher:0.6.0'
-
 }
