+++ app/src/main/java/com/android/news24x7/util/NewsSyncUtils.java
@@ -1,0 +1,83 @@
+package com.android.news24x7.util;
+
+/**
+ * Created by Dell on 6/6/2017.
+ */
+
+import android.content.Context;
+import android.content.Intent;
+import android.database.Cursor;
+import android.support.annotation.NonNull;
+
+import com.android.news24x7.sync.NewsFirebaseJobService;
+import com.android.news24x7.sync.NewsSyncIntentService;
+import com.firebase.jobdispatcher.Constraint;
+import com.firebase.jobdispatcher.Driver;
+import com.firebase.jobdispatcher.FirebaseJobDispatcher;
+import com.firebase.jobdispatcher.GooglePlayDriver;
+import com.firebase.jobdispatcher.Job;
+import com.firebase.jobdispatcher.Lifetime;
+import com.firebase.jobdispatcher.Trigger;
+
+import java.util.concurrent.TimeUnit;
+
+public class NewsSyncUtils {
+
+
+    private static final int REMINDER_INTERVAL_MINUTES = 1;
+    private static final int REMINDER_INTERVAL_SECONDS = (int) (TimeUnit.MINUTES.toSeconds(REMINDER_INTERVAL_MINUTES));
+    private static final int SYNC_FLEXTIME_SECONDS = REMINDER_INTERVAL_SECONDS;
+
+    private static boolean sInitialized;
+    static NewsUtil mNewsUtil;
+    private static final String NEWS_SYNC_TAG = "news-sync";
+
+    static void scheduleFirebaseJobDispatcherSync(@NonNull final Context context) {
+
+        Driver driver = new GooglePlayDriver(context);
+        FirebaseJobDispatcher dispatcher = new FirebaseJobDispatcher(driver);
+
+        /* Create the Job to periodically sync Sunshine */
+        Job syncSunshineJob = dispatcher.newJobBuilder()
+                .setService(NewsFirebaseJobService.class)
+                .setTag(NEWS_SYNC_TAG)
+                .setConstraints(Constraint.ON_ANY_NETWORK)
+                .setLifetime(Lifetime.FOREVER)
+                .setRecurring(true)
+                .setTrigger(Trigger.executionWindow(
+                        REMINDER_INTERVAL_SECONDS,
+                        REMINDER_INTERVAL_SECONDS + SYNC_FLEXTIME_SECONDS))
+                .setReplaceCurrent(true)
+                .build();
+
+        dispatcher.schedule(syncSunshineJob);
+    }
+
+    synchronized public static void initialize(@NonNull final Context context) {
+
+        if (sInitialized) return;
+
+        sInitialized = true;
+        scheduleFirebaseJobDispatcherSync(context);
+        Thread checkForEmpty = new Thread(new Runnable() {
+            @Override
+            public void run() {
+                Cursor cursor = null;
+                 mNewsUtil=new NewsUtil();
+                cursor = mNewsUtil.allNewsCursor(context);
+                if (null == cursor || cursor.getCount() == 0) {
+                    startImmediateSync(context);
+                }
+
+            }
+        });
+
+        checkForEmpty.start();
+    }
+
+    public static void startImmediateSync(@NonNull final Context context) {
+        Intent intentToSyncImmediately = new Intent(context, NewsSyncIntentService.class);
+        context.startService(intentToSyncImmediately);
+    }
+
+}
