+++ app/src/main/java/com/android/news24x7/util/Util.java
@@ -1,5 +1,7 @@
 package com.android.news24x7.util;
 
+import android.content.Context;
+import android.content.Intent;
 import android.support.design.widget.NavigationView;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -8,13 +10,14 @@ import android.support.v4.widget.DrawerLayout;
 import android.view.MenuItem;
 
 import com.android.news24x7.R;
+import com.android.news24x7.activities.SettingsActivity;
 
 /**
  * Created by Dell on 6/2/2017.
  */
 public class Util {
 
-    public static void setupDrawerContent(NavigationView navigationView, final DrawerLayout mDrawerLayout) {
+    public static void setupDrawerContent(NavigationView navigationView, final DrawerLayout mDrawerLayout, final Context context) {
         navigationView.setNavigationItemSelectedListener(
                 new NavigationView.OnNavigationItemSelectedListener() {
                     @Override
@@ -28,6 +31,9 @@ public class Util {
                             case R.id.nav_saved: //do what you want to do;
                                 break;
                             case R.id.nav_settings: //do what you want to do;
+                                context.startActivity(new Intent(context, SettingsActivity.class));
+                                // fragment = AboutFragment.newInstance("About", "arg2");
+                                // Snackbar.make(rootLayout, "HU AM I?!", Snackbar.LENGTH_SHORT).show();
                                 break;
                             case R.id.nav_help: // etc,
                         }
